# OASv3/rippling/RipplingOpenAPI.v1.yaml
openapi: 3.0.0
info:
  title: API Reference
  version: '1.0'
  contact:
    name: Rippling Support
    email: support@rippling.com
  license:
    name: MIT
  description: >-
    Access to Rippling's Employee Management Platform requires an access token
    or API key that is tied to a Rippling client account.


    Please see here for an [overview on the Rippling
    Platform.](https://developer.rippling.com/docs/rippling-api/docs/Getting-Started/a-Introduction.md)


    Additionally, you may find these quick guides useful in getting started.


    - [Getting Started as an Application
    Partner](https://developer.rippling.com/docs/rippling-api/docs/Getting-Started/c-Partners.md)

    - [Getting Started as a Rippling
    Customer](https://rippling.stoplight.io/docs/rippling-api/docs/Getting-Started/b-Customers.md)

    - [Rippling
    Scopes](https://developer.rippling.com/docs/rippling-api/docs/Getting-Started/e-Scopes.md)


    The access token or API key should be used as a bearer token for HTTP
    requests, for instance:


    ```js

    curl --location --request GET
    'https://api.rippling.com/platform/api/companies/current' \

    --header 'Authorization: Bearer REDACTED'

    ```


    ### Using the Interactive Documentation


    Rippling's Documentation Portal allows you to test the API endpoints
    directly within the documentation. To do so, provide your API key or Access
    Token as a header parameter with the form Authorization Bearer: TOKEN.
...
servers:
...
paths:
...
  /departments:
    get:
      summary: GET Departments
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                description: An array of department objects.
                items:
                  $ref: '#/components/schemas/Department'
              examples:
                Success:
                  value:
                    - name: Sales
                      parent: 5ec5595ffa9c4e1728118d6f
                    - name: SMB
                      parent: 5ec5595ffa9c4e1728118d6f
      operationId: get-departments
      description: Returns a list of departments for the specified company.
#      security:
#        - authorizationCode:
#            - company
#            - 'company:departments'
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      tags:
        - A. Companies
  /employees:
    get:
      summary: GET Employees
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
              examples:
                example-1:
                  value:
                    - id: string
                      name: string
                      firstName: string
                      lastName: string
                      employmentType: CONTRACTOR
                      title: string
                      gender: MALE
                      identifiedGender: MALE
                      department: string
                      workLocation:
                        city: string
                        streetLine1: string
                        zip: string
                        country: string
                        state: string
                        streetLine2: string
                      spokeId: string
                      endDate: string
                      roleState: INIT
                      workEmail: string
                      manager: string
                      customFields:
                        Id: string
                        type: TEXT
                        title: string
                        required: true
      operationId: get-employees
#      security:
#        - authorizationCode:
#            - employee
      description: >-
        Retrieves the list of active employees. The fields retrieved depend on
        the employee scopes that you have access to for your access token or API
        key. The only guarenteeed fields include id, personalEmail, and
        roleState.
      tags:
        - B. Employees
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
  /work_locations:
    get:
      summary: GET Work Locations
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkLocation'
              examples:
                example-1:
                  value:
                    - nickname: string
                      address:
                        city: string
                        streetLine1: string
                        zip: string
                        country: string
                        state: string
                        streetLine2: string
      operationId: get-work_locations
      description: Returns the list of work locations for a given company.
#      security:
#        - authorizationCode:
#            - company
#            - 'company:workLocations'
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      tags:
        - A. Companies
    parameters: []
...
  /teams:
    get:
      summary: GET Teams
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
              examples:
                Success:
                  value:
                    - name: Applications
                      id: 5ee6c69bbc3bc7293a217ff2
                      parent: null
                    - name: Identity
                      id: 5ee6c6a2fa9c4e3d71cc32d7
                      parent: null
                    - name: Hardware
                      id: 5ee6c6a7bc3bc7293a217ff6
                      parent: null
                    - name: Payroll
                      id: 5ee6c6abbc3bc72c3f6b88ef
                      parent: null
      operationId: get-teams
      description: Retrieves the list of teams for the company.
#      security:
#        - authorizationCode:
#            - company
#            - 'company:teams'
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      tags:
        - A. Companies
...
  /employees/include_terminated:
    get:
      summary: GET Employees (Including Terminated)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
              examples:
                example-1:
                  value:
                    - id: string
                      name: string
                      firstName: string
                      lastName: string
                      employmentType: CONTRACTOR
                      title: string
                      gender: MALE
                      identifiedGender: MALE
                      department: string
                      workLocation:
                        city: string
                        streetLine1: string
                        zip: string
                        country: string
                        state: string
                        streetLine2: string
                      spokeId: string
                      endDate: string
                      roleState: INIT
                      workEmail: string
                      manager: string
                      customFields:
                        Id: string
                        type: TEXT
                        title: string
                        required: true
      operationId: get-employees-include_terminated
      description: >-
        Similar to the active employees endpoint, but this endpoint includes
        both active and terminated employees.
#      security:
#        - authorizationCode:
#            - employee
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/ein'
      tags:
        - B. Employees
...
  /groups:
    get:
      summary: GET Groups
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
              examples:
                Success:
                  value:
                    - id: 5f397ed33ca04e4ab23246c2
                      spokeId: 123456xvbnm12345678912354939
                      name: GroupTest3
                      users: []
                      version: t4lz8cjhc54ludro
                    - id: 5f397f3b3ca04e478d4575ad
                      spokeId: lkjwbfskjsdbllskdb
                      name: GroupTest4
                      users:
                        - 5c8f7f06c592917aeee1ea9f
                      version: uu8ccavwbsb8stfn
      operationId: get-groups
      description: >-
        Please note, the Groups endpoint requires an OAuth application (i.e.
        approved 3rd party partners), as the end point is intended for mapping
        third-party application “Groups” within Rippling organizations.


        Lists the current third-party groups for an organization.
#      security:
#        - authorizationCode:
#            - 'group_members:read'
#            - 'groups:read'
      tags:
        - F. Groups
    parameters: []
    post:
      summary: POST Groups
      operationId: post-groups
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
              examples:
                Success:
                  value:
                    id: 5f397f3b3ca04e478d4575ad
                    spokeId: lkjwbfskjsdbllskdb
                    name: GroupTest4
                    users:
                      - 5c8f7f06c592917aeee1ea9f
                    version: uu8ccavwbsb8stfn
      description: >-
        Creates a generic group, that can be associated within the third-party
        application.
#      security:
#        - authorizationCode:
#            - 'groups:write'
#            - 'group_members:write'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: User-readable name of the group.
                spokeId:
                  type: string
                  description: >-
                    The unique ID for the group, this can be the unique
                    identifier for the group entity object within your
                    application.
                users:
                  type: array
                  description: An array of Rippling IDs that will be in the group.
                  items:
                    type: string
            examples:
              New Group:
                value:
                  name: GroupTest4
                  spokeId: lkjwbfskjsdbllskdb
                  users:
                    - 5c8f7f06c592917aeee1ea9f
        description: A group object.
      tags:
        - F. Groups
...
  /leave_requests:
    get:
      summary: GET Leave Requests
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaveRequest'
              examples:
                Success:
                  value:
                    - id: 5b998dfb40cae798a83e8874
                      createdAt: '2018-08-12T15:06:51.681000-07:00'
                      updatedAt: '2018-08-13T15:06:51.684000-07:00'
                      role: 5ae0b2cb40cae73202bd66cc
                      roleName: Jane Doe
                      requestedBy: 5ae0b2cb40cae73202bd66cc
                      requestedByName: Jane Doe
                      status: APPROVED
                      startDate: '2018-08-22'
                      endDate: '2018-08-24'
                      comments: null
                      numHours: 20
                      numMinutes: 1200
                      leavePolicy: 5b998cf940cae798a83e87fd
                      leaveType: VACATION
                      policyDisplayName: Vacation
                      processedAt: '2018-08-13T15:06:51.684000-07:00'
                      processedBy: 5ae0b2cb40cae73202bd66cc
                      processedByName: Funky Admin
                      reasonForLeave: Visiting hometown for my brother’s wedding
                      dates:
                        - date: '2018-08-22'
                          numMinutes: 480
                        - date: '2018-08-23'
                          numMinutes: 480
                        - date: '2018-08-24'
                          numMinutes: 240
      operationId: get-leave_requests
      description: >-
        Retrieves the current leave requests.The query can be filtered by a
        number of specific query parameters.
      parameters:
        - schema:
            type: string
          in: query
          name: id
        - schema:
            type: string
          in: query
          name: role
        - schema:
            type: string
          in: query
          name: requestedBy
        - schema:
            type: string
          in: query
          name: status
        - schema:
            type: string
          in: query
          name: startDate
          description: Start date of leave.
        - schema:
            type: string
          in: query
          name: endDate
          description: End date of leave.
        - schema:
            type: string
          in: query
          name: leavePolicy
        - schema:
            type: string
          in: query
          name: processedBy
        - schema:
            type: string
          in: query
          name: from
          description: >-
            Filter to capture whether the leave request overlaps with a date
            range.
        - schema:
            type: string
          in: query
          name: to
          description: >-
            Filter to capture whether the leave request overlaps with a date
            range.
#      security:
#        - authorizationCode:
#            - company
#            - 'company:leave_requests'
      tags:
        - A. Companies
...
  /ats_candidates/push_candidate:
    post:
      summary: POST New Candidate
      operationId: post-ats_candidates-push_candidate
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
      description: >-
        Pushes a candidate from an applicant tracking system directly into the
        Rippling onboarding flow. Please note, this endpoint is only available
        to applications integrating with OAuth2.0.
#      security:
#        - authorizationCode:
#            - 'ats:candidates'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
        description: ''
      tags:
        - H. ATS
...
components:
...
  securitySchemes:
    Bearer ApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: >-
        If you are using Rippling's API to access endpoints on behalf of your
        own company, please use your API key. Your API key can be retrieved from
        the Company Settings page on Rippling.
    authorizationCode:
      type: oauth2
      flows:
        authorizationCode:
          tokenUrl: 'https://app.rippling.com/api/o/token/'
          scopes:
            company: >-
              The parent company scope. This is a prequisite to additional
              company:* scopes, but does not serve as a scope superset. 
            employee: >-
              The parent employee scope. This is a prerequisite to additional
              employee:* scopes, but does not serve as a scope superset. 
            'company:activity': Permission to access information regarding the company's activity.
            'company:address': Permission to access the company's address.
            'company:phone': Permission to access the company's phone information.
            'company:departments': Permission to access the company's departments.
            'company:levels': Permission to access the company's levels.
            'company:customFields': Permission to access the company's custom fields.
            'company:workLocations': Permission to access the company's work locations.
            'company:leave_requests': Permission to access the company's leave request information.
            'admin:name': Permission to access the company's admins' names.
            'admin:personalEmail': Permission to access the company's admins' personal emails.
            'admin:phone': Permission to access the company's admins' phone number.
            'admin:workEmail': Permission to access the company's admins' work emails.
            'admin:roleState': Permission to access the company's admins' role states.
            'admin:department': Permission to access the company's admins' departments.
            'admin:workLocation': Permission to access the company's admins' work location.
            'employee:name': Permission to access the company's employees' names.
            'employee:employmentType': Permission to access the company's employees' employment types.
            'employee:workEmail': Permission to access the company's employees' work emails.
            'employee:title': Permission to access the company's employees' titles.
            'employee:endDate': Permission to access the company's employees' end dates.
...